+++ c/proxy/conf/nginx.conf
@@ -4,7 +4,8 @@ worker_processes  2;
 
 #error_log  logs/error.log;
 #error_log  logs/error.log  notice;
-error_log  logs/error.log  info;
+error_log  logs/info.log  info;
+error_log  logs/error.log  error;
 
 #pid        logs/nginx.pid;
