+++ c/proxy/conf/route_table.yaml
@@ -1,22 +1,27 @@
 default:
   proxy:
     - host: 127.0.0.1
-      port: 9390
+      http_port: 9300
+      grpc_port: 9310
 local:
   test_proxy:
     - host: 127.0.0.1
-      port: 9302
+      http_port: 9302
 9999:
   proxy:
     - host: 127.0.0.1
-      port: 9390
+      http_port: 9300
+      grpc_port: 9310
   fateflow:
     - host: 127.0.0.1
-      port: 9360
+      http_port: 9380
+      grpc_port: 9360
 10000:
   proxy:
     - host: 127.0.0.1
-      port: 9390
+      http_port: 9300
+      grpc_port: 9310
   fateflow:
     - host: 127.0.0.1
-      port: 9360
+      http_port: 9380
+      grpc_port: 9360
