+++ c/proxy/conf/vhost/coordination_grpc_proxy.conf
@@ -1,5 +1,5 @@
 server {
-    listen       9390 http2;
+    listen       9310 http2;
     server_name  coordination_grpc_proxy;
 
     #charset koi8-r;
